requirements.txt
python-dotenv
requests
beautifulsoup4
lxml
feedparser
pandas
langdetect
Sastrawi
tqdm
